Update of all random routines; semantic clarified; own versions for sparse fill, update doc.

+++ scai/lama/matutils/MatrixCreator.hpp
@@ -65,7 +65,7 @@ public:
      *
      */
 
-    static void fillRandom( Matrix& matrix, double density );
+    static void fillRandom( Matrix& matrix, float density );
 
     /** Builds a block distributed matrix with random values.
      *
